dlbot-lambda/requirements.txt
boto3
yt-dlp
mutagen
python-telegram-bot